Rename package from `isr-spectrum`/`isr_spectrum` to `inscar`
