+++ docs/modules.rst
@@ -4,4 +4,4 @@ Modules
 .. toctree::
    :maxdepth: 4
 
-   isr_spectrum
+   inscar
