+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "isr-spectrum"
+name = "inscar"
 version = "3.0.0"
 description = "Calculate an incoherent scatter spectrum with arbitrary isotropic electron velocity distributions and radar pointing at oblique angles to the magnetic field"
 authors = ["engeir <eirroleng@gmail.com>"]
@@ -20,9 +20,7 @@ tqdm = "^4.59.0"
 si-prefix = "^1.2.2"
 isort = "^5.8.0"
 flake8 = "^3.9.0"
-pre-commit = "^2.11.1"
 pytest = "^6.2.2"
-pre-commit-hooks = "^3.4.0"
 pytest-cov = "^3.0.0"
 black = "^22.3.0"
 darglint = "^1.8.1"
@@ -33,6 +31,8 @@ mypy = "^0.950"
 Sphinx = "^4.5.0"
 sphinx-autobuild = "^2021.3.14"
 sphinx-rtd-theme = "^1.0.0"
+pre-commit = "^2.19.0"
+pre-commit-hooks = "^4.2.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
