+++ src/inscar/numba_integration.py
@@ -7,7 +7,7 @@ import numba as nb
 import numpy as np
 import scipy.constants as const
 
-import isr_spectrum as isr
+import inscar as isr
 
 
 @nb.njit(parallel=True, cache=True)
